configure resolver via hooksets
