+++ atlas/scaife_viewer/atlas/resolvers/common.py
@@ -1,0 +1,5 @@
+class Library:
+    def __init__(self, text_groups, works, versions):
+        self.text_groups = text_groups
+        self.works = works
+        self.versions = versions
